+++ tgbotapi.core/src/jsMain/kotlin/dev/inmo/tgbotapi/utils/DefaultKtorEngine.js.kt
@@ -1,0 +1,8 @@
+package dev.inmo.tgbotapi.utils
+
+import io.ktor.client.engine.HttpClientEngineFactory
+import io.ktor.client.engine.js.Js
+
+actual val defaultKtorEngine: HttpClientEngineFactory<*> by lazy {
+    Js
+}