+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 28.0.3
 
+* `Core`:
+  * Add passing of default engines in `HttpClient` constructors
+
 ## 28.0.2
 
 * `Core`: